+++ mch/core_v3/auth.go
@@ -1,13 +1,18 @@
 package core_v3
 
 import (
+	"crypto"
 	"crypto/aes"
 	"crypto/cipher"
+	"crypto/rsa"
+	"crypto/sha256"
 	"crypto/x509"
 	"encoding/base64"
 	"encoding/json"
 	"encoding/pem"
 	"fmt"
+	"math/rand"
+	"time"
 )
 
 type CertResponse struct {
@@ -83,5 +88,49 @@ func (cli *Client) GetWechatCertificate() (err error) {
 	cli.WechatCertificate = cert
 	return
 }
+func (cli *Client) Sign(method, url, body, nonce string) (sign string, err error) {
+	var string2sign string
+	var digest [32]byte
+	var hashed []byte
 
-func main() {}
+	string2sign = method + "\n" + url + "\n" + timestamp() + "\n" + nonce + "\n" + body + "\n"
+	if DEBUG {
+		fmt.Println("string to sign: ", string2sign)
+	}
+
+	digest = sha256.Sum256([]byte(string2sign))
+	hashed, err = rsa.SignPKCS1v15(nil, cli.PrivateKey, crypto.SHA256, digest[:])
+	if err != nil {
+		return
+	}
+
+	sign = base64.StdEncoding.EncodeToString(hashed)
+
+	return
+}
+
+func genRandomString(size int) string {
+
+	if DEBUG {
+		return "123456"
+	}
+
+	var a, b []byte
+	var i int
+
+	a = []byte("1234567890abcdefghijklmnopqrstuvwxyz")
+	b = make([]byte, size)
+
+	for i = 0; i < size; i++ {
+		b[i] = a[rand.Intn(36)]
+	}
+
+	return string(b)
+}
+
+func timestamp() string {
+	// if DEBUG {
+	// 	return "1582883827"
+	// }
+	return fmt.Sprintf("%d", time.Now().Unix())
+}
