+++ mch/core_v3/client_test.go
@@ -35,15 +35,15 @@ func Test_DoGet(t *testing.T) {
 	fmt.Println(string(resp))
 }
 
-func Test_GetWechatCert(t *testing.T) {
-	fmt.Println("----------------------------------")
-	fmt.Println("Testing get wechat certificate")
+// func Test_GetWechatCert(t *testing.T) {
+// 	fmt.Println("----------------------------------")
+// 	fmt.Println("Testing get wechat certificate")
 
-	var err error
-	var cli *Client
+// 	var err error
+// 	var cli *Client
 
-	cli, _ = NewClient("1533391551", "72E6A550FCCC4AB90E1699D06989669221DF167A", "001rsrs001001rsrs001001rsrs001nb", "./apiclient_cert.pem", "./apiclient_key.pem")
+// 	cli, _ = NewClient("1533391551", "72E6A550FCCC4AB90E1699D06989669221DF167A", "001rsrs001001rsrs001001rsrs001nb", "./apiclient_cert.pem", "./apiclient_key.pem")
 
-	err = cli.GetWechatCertificate()
-	assertEqual(t, err, nil)
-}
+// 	err = cli.GetWechatCertificate()
+// 	assertEqual(t, err, nil)
+// }
